add unit tests for ProductService

+++ src/modules/products/product.service.ts
@@ -13,9 +13,11 @@ export class ProductService {
   }
 
   async getProduct(id: number): Promise<Product | null> {
-    if (id <= 0 || id === null) throw new Error('Id inválido');
+    if (id <= 0) throw new Error('Id inválido');
 
     const product = await this.repo.getById(id);
+
+    if (!product) throw new Error('Produto não existe');
 
     return product;
   }
@@ -35,7 +37,7 @@ export class ProductService {
 
     const existing = await this.repo.getById(id);
 
-    if (!existing) throw new Error('Produto Não existe');
+    if (!existing) throw new Error('Produto não existe');
 
     const updateProduct = await this.repo.updateById(id, data);
 
@@ -45,7 +47,11 @@ export class ProductService {
   async removeProduct(id: number): Promise<Product | null> {
     if (id <= 0) throw new Error('Id inválido');
 
+    const existing = await this.repo.getById(id);
+    if (!existing) throw new Error('Produto não existe');
+
     const product = await this.repo.deleteById(id);
+
     return product;
   }
 }
